Condense all hooks to a single file
