--- src/app/Utils/hooks/useAppDispatch.ts
@@ -1,4 +1,0 @@
-import { useDispatch } from "react-redux";
-import { AppDispatch } from "../../Store/store";
-
-export const useAppDispatch = () => useDispatch<AppDispatch>();
