--- src/app/Utils/hooks/useCartState.ts
@@ -1,8 +1,0 @@
-/** @format */
-
-import { RootState } from "../../Store/store";
-import { useAppSelector } from "./useAppSelector";
-
-
-export const useCartState = () =>
-  useAppSelector((state: RootState) => state.car.cart);
